core-*, stress-*: use pid_t for waitpid() returns and add PID for errors
More portability fixes, used pid_t for all waitpid returns rather than
int. Also print the PID when reporting waitpid errors.

+++ test/test-waitpid.c
@@ -24,5 +24,5 @@ int main(void)
 {
 	int status;
 
-	return waitpid(999999, &status, 0);
+	return (int)waitpid(999999, &status, 0);
 }
